projects/rawkode-social/video-to-post/requirements.txt
deepgram-sdk
ffmpy
marimo
yt-dlp
